+++ fapi/main.py
@@ -9,7 +9,7 @@ from fapi.api.routes import (
     candidate, leads, google_auth, talent_search, user_role,
     contact, login, register, resources, vendor_contact,
     vendor, vendor_activity, request_demo, unsubscribe,
-    user_dashboard, password,employee,dashboard,
+    user_dashboard, password,employee,avatar_dashboard,
 )
 from fapi.core.config import limiter  
 
@@ -36,7 +36,7 @@ app.include_router(resources.router, prefix="/api", tags=["Resources"])
 app.include_router(register.router, prefix="/api", tags=["Register"])
 app.include_router(request_demo.router, prefix="/api", tags=["Request Demo"])
 app.include_router(user_dashboard.router, prefix="/api", tags=["User Dashboard"])
-app.include_router(dashboard.router, prefix="/api", tags=["Dashboard"])
+app.include_router(avatar_dashboard.router, prefix="/api", tags=["Avatar Dashboard"])
 
 
 def get_db():
